PlumeHandler class

Internal class processing and shaping tabular data into a plume object.

Super class

plume::NameHandler -> PlumeHandler

Methods

Public methods

PlumeHandler$new()

PlumeHandler$print()

PlumeHandler$get_plume()

PlumeHandler$get_roles()

PlumeHandler$clone()

Method new()

Usage

PlumeHandler$new(
  data,
  names,
  roles,
  credit_roles,
  initials_given_name,
  family_name_first = FALSE,
  interword_spacing = TRUE
)

Method print()

Usage

PlumeHandler$print()

Method get_plume()

Usage

PlumeHandler$get_plume()

Method get_roles()

Usage

PlumeHandler$get_roles()

Method clone()

The objects of this class are cloneable with this method.

Usage

PlumeHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.