Push plume objects into a Quarto document

plm_push() inserts author information not handled by Quarto's author schema into a Quarto document. This function should only be used with PlumeQuarto to inject author data from a separate script that are not passed via PlumeQuarto$to_yaml(). It's handy to avoid instantiating the class both in the R script and Quarto file.

Usage

plm_push(x, file, where = NULL, sep = "; ")

Arguments

x

A plume object.

file

A .qmd file.

where

Line of text present in file after which data should appear. This parameter is ignored if an anchor for the specific data to insert already exits in the file.

sep

Separator used to separate items passed to x.

Value

The input file invisibly.

Examples

if (FALSE) {
tmp_file <- withr::local_tempfile(
  lines = "# Foo\n\n# Bar",
  fileext = ".qmd"
)

readr::read_file(tmp_file) |> cat()

o <- PlumeQuarto$new(
  encyclopedists,
  names = c(contribution = "contribution_n")
)
o$get_contributions() |> plm_push(tmp_file, where = "# Foo")

readr::read_file(tmp_file) |> cat()
}