plume: A Simple Author Handler for Scientific Writing

Handles and formats author information in scientific writing in 'R Markdown' and 'Quarto'. 'plume' provides easy-to-use and flexible tools for injecting author metadata in 'YAML' headers as well as generating author and contribution lists (among others) as strings from tabular data.

See also

Author

Maintainer: Arnaud Gallou arangacas@gmail.com (ORCID) [copyright holder]