PlumeQuarto class

Class that pushes or updates author metadata in Quarto files.

Super classes

plume::NameHandler -> plume::PlumeHandler -> plume::StatusSetter -> plume::StatusSetterQuarto -> PlumeQuarto

Methods

Public methods

PlumeQuarto$new()

PlumeQuarto$to_yaml()

PlumeQuarto$clone()

Inherited methods

plume::PlumeHandler$get_plume()

plume::PlumeHandler$print()

plume::StatusSetter$set_corresponding_authors()

plume::StatusSetterQuarto$set_deceased()

plume::StatusSetterQuarto$set_equal_contributor()

Method new()

Create a PlumeQuarto object.

Usage

PlumeQuarto$new(
  data,
  names = NULL,
  credit_roles = FALSE,
  initials_given_name = FALSE,
  by = NULL
)

Arguments

data

A data frame containing author-related data.

names

A vector of key-value pairs specifying custom names to use, where keys are default names and values their respective replacements.

credit_roles

Should the Contributor Roles Taxonomy be used?

initials_given_name

Should the initials of given names be used?

by

A character string defining the default variable used to assign authors' status in all set_*() methods. By default, uses authors' id.

Returns

A PlumeQuarto object.

Method to_yaml()

Push or update author information in a YAML header. The generated YAML complies with Quarto'author and affiliations schemas)`.

Usage

PlumeQuarto$to_yaml(file)

Arguments

file

A .qmd file.

Details

If missing, to_yaml() pushes author information into a YAML header. If already existing, the function replaces old author and affiliations values with the ones provided in the input data.

Returns

The input file invisibly.

Method clone()

The objects of this class are cloneable with this method.

Usage

PlumeQuarto$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.