StatusSetter class

Internal class that manages authors' status.

Super classes

plume::NameHandler -> plume::PlumeHandler -> StatusSetter

Methods

Public methods

StatusSetter$new()

StatusSetter$set_corresponding_authors()

StatusSetter$clone()

Inherited methods

plume::PlumeHandler$get_plume()

plume::PlumeHandler$get_roles()

plume::PlumeHandler$print()

Method new()

Usage

StatusSetter$new(..., by)

Method set_corresponding_authors()

Set corresponding authors.

Usage

StatusSetter$set_corresponding_authors(..., .by, by = deprecated())

Arguments

...

One or more unquoted expressions separated by commas. Expressions matching values in the column defined by by/.by determine corresponding authors. Matching of values is case- insensitive.

.by

Variable used to set corresponding authors. By default, uses authors' id.

by

Please use the .by parameter instead.

Returns

The class instance.

Method clone()

The objects of this class are cloneable with this method.

Usage

StatusSetter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.