StatusSetterPlumeQuarto class

Internal class extending StatusSetter for PlumeQuarto.

Super classes

plume::NameHandler -> plume::PlumeHandler -> plume::StatusSetter -> StatusSetterPlumeQuarto

Methods

Public methods

StatusSetterPlumeQuarto$set_cofirst_authors()

StatusSetterPlumeQuarto$set_equal_contributor()

StatusSetterPlumeQuarto$set_deceased()

StatusSetterPlumeQuarto$clone()

Inherited methods

plume::PlumeHandler$get_plume()

plume::PlumeHandler$get_roles()

plume::PlumeHandler$print()

plume::StatusSetter$initialize()

plume::StatusSetter$set_corresponding_authors()

Method set_cofirst_authors()

Set co-first authors.

Usage

StatusSetterPlumeQuarto$set_cofirst_authors(..., .by)

Arguments

...

One or more unquoted expressions separated by commas. Expressions matching values in the column defined by by/.by determine co-first authors. Matching of values is case-insensitive.

.by

Variable used to specify which authors contributed equally to the work. By default, uses authors' id.

Returns

The class instance.

Method set_equal_contributor()

This method has been deprecated in favour of set_cofirst_authors().

Usage

StatusSetterPlumeQuarto$set_equal_contributor(..., .by, by = deprecated())

Arguments

...

One or more unquoted expressions separated by commas. Expressions matching values in the column defined by by/.by determine equal contributors. Matching of values is case-insensitive.

.by

Variable used to specify which authors are equal contributors. By default, uses authors' id.

by

Please use the .by parameter instead.

Returns

The class instance.

Method set_deceased()

Set deceased authors.

Usage

StatusSetterPlumeQuarto$set_deceased(..., .by, by = deprecated())

Arguments

...

One or more unquoted expressions separated by commas. Expressions matching values in the column defined by by/.by determine deceased authors. Matching of values is case-insensitive.

.by

Variable used to specify whether an author is deceased or not. By default, uses authors' id.

by

Please use the .by parameter instead.

Returns

The class instance.

Method clone()

The objects of this class are cloneable with this method.

Usage

StatusSetterPlumeQuarto$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.